Turn the entire schedule editor form into angular
This is still not completely finished, as there is still a bit to clean
up in the controller. The mechanism to detect whether or not there have
been changes to the form has also not been implemented in this commit,
and in fact the buttons remain unclickable even when changing things.
This commit is already large, so I wanted to make an attempt to break it
up.